Banner Login/Signup Popup
I don't remember everything, but this includes the items and changes to
have the login and signup as popups within the top banner via Ajax calls

-This includes a BannerServlet that handles the ajax request from the
banner, such as login/signup and template pulls.

-A banner.ftl so that you can call the template before your other
temaplate calls to load in the banner(header) of the page and not have
the entire banner on the top of every page. It also allows you to pass
it a list called 'scripts' in the root map and it will load those
scripts into the page.

-A Banner.js to handle the popup feature. Ajax call to request the
specific template that it wants/needs.

-An updated login and signup ftls and scripts.