Added sign out functionality as well as account/banner popup
-Added an account popup ftl with "My account","My bookings","admin
page", and "sign out" buttons.

-The banner now detects if a session is already there by checking if
there is a Java 'user' object attributed to it.

-For now it just passes the user info via a user.toString method, but it
probably should be JSON. I just don't know how to work with JSON and I
was testing the session stuff

-the login and sign up scripts are now one. "loginSignup.js"

-added a method to loginLogic to pull the first/last name from the user,
so that the front end can show it on the account page and banner. It
takes in the user id and returns a string array of size 2. {first,last}